Altered formatting and file hierarchy as per code review
