+++ keyboards/handwired/bongo80/bongo80.h
@@ -1,4 +1,4 @@
-/* Copyright 2023 sam kelly (@samkellu)
+/* Copyright 2023 Sam Kelly (@samkellu)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -12,13 +12,10 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
+ */
 
 #pragma once
 
-#include_next <mcuconf.h>
-
-#undef RP_I2C_USE_I2C0
-#define RP_I2C_USE_I2C0 TRUE
-#undef RP_I2C_USE_I2C1
-#define RP_I2C_USE_I2C1 FALSE
+enum keycodes {
+	KC_OLED_STATE = SAFE_RANGE,
+};
