--- keyboards/handwired/bongo80/rules.mk
@@ -1,1 +1,0 @@
-SRC += bongo80.c
